Smack: bidirectional UDS connect check
Smack IPC policy requires that the sender have write access
to the receiver. UDS streams don't do per-packet checks. The
only check is done at connect time. The existing code checks
if the connecting process can write to the other, but not the
other way around. This change adds a check that the other end
can write to the connecting process.

Targeted for git://git.gitorious.org/smack-next/kernel.git

Change-Id: Ibecf37e5644f66885acc02ae0ee96f8009a0b5fa
Signed-off-by: Casey Schuafler <casey@schaufler-ca.com>
[Backport from mainline to 3.10]
Signed-off-by: Seung-Woo Kim <sw0312.kim@samsung.com>